removed hardcoded giles path
