+++ conf/net/int_service/giles_conf.json
@@ -1,0 +1,4 @@
+{
+	"giles_base_url": "http://localhost:8079",
+	"giles_api_key": "fakeapikey"
+}
